+++ dev_tools/test_alconna_2.py
@@ -2,7 +2,7 @@ from typing import Union, Dict
 
 from arclet.alconna import compile
 from arclet.alconna.builtin.construct import AlconnaString, AlconnaFormat
-from arclet.alconna.typing import UnionArg, pattern_gen
+from arclet.alconna.typing import pattern_gen
 from arclet.alconna import Alconna, Args, Option
 from arclet.alconna import command_manager
 from graia.ariadne.message.element import At
@@ -10,13 +10,13 @@ from devtools import debug
 
 print(command_manager)
 
-ping = "Test" @ AlconnaString("ping <url:url>")
+ping = "Test" / AlconnaString("ping <url:url>")
 ping1 = AlconnaString("ping <url:url>")
 
 Alconna.set_custom_types(digit=int)
 alc = AlconnaFormat(
     "lp user {target} perm set {perm} {default}",
-    {"target": str, "perm": str, "default": Args["de":bool:True]},
+    {"target": str, "perm": str, "default": Args["de", bool, True]},
 )
 alcc = AlconnaFormat(
     "lp1 user {target}",
@@ -24,7 +24,7 @@ alcc = AlconnaFormat(
 )
 
 alcf = AlconnaFormat("music {artist} {title:str} singer {name:str}")
-print(alcf.parse("music --help"))
+alcf.parse("music --help")
 debug(alc)
 alc.exception_in_time = False
 debug(alc.parse("lp user AAA perm set admin"))
@@ -53,14 +53,15 @@ def test_type(text: str):
     return eval(text)
 
 
-alc2 = Alconna("test", help_text="测试help直接发送") + Option("foo", Args["bar":str, "bar1":int:12345, "bar2":test_type])
-print(alc2.parse("test --help"))
+alc2 = Alconna("test", help_text="测试help直接发送") + \
+       Option("foo", Args["bar", str]["bar1", int, 12345]["bar2", test_type])
+alc2.parse("test --help")
 
 alc4 = Alconna(
     command="test_multi",
     options=[
-        Option("--foo", Args["tags;S":str:1, "str1":str]),
-        Option("--bar", Args["num": int]),
+        Option("--foo", Args["tags;S", str, 1]["str1", str]),
+        Option("--bar", Args["num", int]),
     ]
 )
 
@@ -73,7 +74,7 @@ print(result.query("foo"))
 alc5 = Alconna("test_anti", "path;A:int")
 print(alc5.parse("test_anti a"))
 
-alc6 = Alconna("test_union", main_args=Args.path[UnionArg[int, float, 'abc']])
+alc6 = Alconna("test_union", main_args=Args.path[[int, float, 'abc']])
 print(alc6.parse("test_union abc"))
 print(alc6.parse(["test_union 123"]))
 
@@ -96,10 +97,10 @@ print(alc10.get_help())
 alc11 = Alconna("test_header", headers=[(At(123456), "abc")])
 print("alc11:", alc11.parse([At(123456), "abctest_header"]))
 
-alc12 = Alconna("test_str1", Args["abcd", "1234"])
+alc12 = Alconna("test_str1", Args["abcd"]["1234", "1234"])
 print("alc12:", alc12.parse("test_str1 abcd 1234"))
 
-alc13 = Alconna("image", Args["--width;O|K":int:1920, "--height;O|K":int:1080])
+alc13 = Alconna("image", Args["--width;O|K", int, 1920]["--height;O|K", int, 1080])
 print("alc13:", alc13.parse("image --height=720"))
 
 alc14 = Alconna(main_args="foo:str", headers=['!test_fuzzy'], is_fuzzy_match=True)
@@ -110,8 +111,8 @@ print(alc15.parse("my_string --foo 123 --bar"))
 
 alc16 = Alconna(
     "发涩图",
-    Args["min":r"(\d+)张", "max;O":r"(\d+)张"] / "到",
-    options=[Option("从", Args["tags;3":str] / "和")],
+    Args["min", r"(\d+)张"]["max;O", r"(\d+)张"] / "到",
+    options=[Option("从", Args["tags;3", str] / "和")],
     action=lambda x, y: (int(x), int(y))
 )
 s_alc15 = compile(alc15)